+++ test/doctests.hs
@@ -3,4 +3,4 @@ module Main where
 import           Test.DocTest
 
 main :: IO ()
-main = doctest ["-isrc", "src/Test/Hspec/Util.hs"]
+main = doctest ["-isrc", "-optP-include", "-optPdist/build/autogen/cabal_macros.h", "src/Test/Hspec/Util.hs", "src/Test/Hspec/Formatters.hs"]